WEBAPP-333: Fix tests
